Updated post
Hi @ivelasq - I have added the changes to the blog. I have created sections on evaluating the model and splitting the training data, other than that, you may want to link in the article I sent to you on interpreting logistic regression outputs. So p-value and odds plots are two methods, but there are heaps. I like these, because they are nice and visual. Furthermore, did a quality check of the odds ratios returned from the parsnip model fit object by taking the exponent of the estimates. This aligns 100% with the OddsPlotty package. Thanks